leetcode Add Strings

+++ src/leetcode/AddStrings.java
@@ -14,6 +14,6 @@ public class AddStrings {
 
     private static String addStrings(String num1, String num2) {
         var sum = Double.parseDouble(num1) + Double.parseDouble(num2);
-        return String.valueOf(sum);
+        return String.valueOf((long) sum);
     }
 }
